+++ app/blog/page.tsx
@@ -2,6 +2,7 @@ import fs from "fs";
 import path from "path";
 import matter from "gray-matter";
 import Link from "next/link";
+import Layout from "@/src/layout/index";
 import Title from "@/src/components/Title";
 import LineBreak from "@/src/components/LineBreak";
 
@@ -22,7 +23,7 @@ export default function Home() {
     .sort((a, b) => b.meta.id - a.meta.id); 
 
   return (
-      <>
+    <Layout>
       <Title>Blog</Title>
       <LineBreak />
       <div>
@@ -42,7 +43,7 @@ export default function Home() {
           </Link>
         ))}
       </div>
-      </>
+    </Layout>
   );
 }
