+++ app/globals.css
@@ -66,7 +66,6 @@
   }
 } */
 
-
 @layer base {
   :root {
     --background: 0 0% 100%;
@@ -88,7 +87,7 @@
     --border: 240 5.9% 90%;
     --input: 240 5.9% 90%;
     --ring: 240 5.9% 10%;
-    --radius: 0rem;
+    --radius: 0.5rem;
   }
 
   .dark {
