+++ app/not-found.tsx
@@ -1,14 +1,15 @@
 "use client";
 
+import Layout from "@/src/layout/index";
 import { MDXProvider } from "@mdx-js/react";
 import NotFound from "@/src/content/404.mdx";
 
 export default function page() {
   return (
-      <>
-        <MDXProvider components={{}}>
+    <Layout>
+      <MDXProvider components={{}}>
         <NotFound />
       </MDXProvider>
-      </>
+    </Layout>
   );
 }
