+++ app/page.tsx
@@ -1,15 +1,15 @@
 "use client";
 
+import Layout from "@/src/layout/index";
 import { MDXProvider } from "@mdx-js/react";
 import Iam from "@/src/content/iam.mdx";
 
 export default function page() {
   return (
-    
-      <>
+    <Layout>
       <MDXProvider components={{}}>
         <Iam />
       </MDXProvider>
-      </>
+    </Layout>
   );
 }
