+++ app/projects/page.tsx
@@ -1,14 +1,15 @@
 "use client";
 
+import Layout from "@/src/layout/index";
 import { MDXProvider } from "@mdx-js/react";
 import Projects from "@/src/content/projects.mdx";
 
 export default function page() {
   return (
-     <>
-       <MDXProvider components={{}}>
+    <Layout>
+      <MDXProvider components={{}}>
         <Projects />
       </MDXProvider>
-     </>
+    </Layout>
   );
 }
