+++ app/secretbook/page.tsx
@@ -1,14 +1,15 @@
 "use client";
 
+import Layout from "@/src/layout/index";
 import { MDXProvider } from "@mdx-js/react";
 import SecretBook from "@/src/content/secretbook.mdx";
 
 export default function page() {
   return (
-      <>
-        <MDXProvider components={{}}>
+    <Layout>
+      <MDXProvider components={{}}>
         <SecretBook />
       </MDXProvider>
-      </>
+    </Layout>
   );
 }
